+++ Runtime/System/Splines/NCurveBehaviour.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Nazio_LT.Tools.Core
@@ -8,7 +7,13 @@ namespace Nazio_LT.Tools.Core
     {
         //For Editor
         public bool editing;
+        [SerializeField, ] public float dist;
 
         [SerializeField] public NCurve curve;
+
+        private void OnDrawGizmos()
+        {
+            Gizmos.DrawCube(curve.ComputePointDistance(dist), Vector3.one * 0.5f);
+        }
     }
 }