+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 * Color set alpha method. ([f8d247b](https://github.com/NazioLT/NZCore/commit/f8d247b91c5566ee6a3cb4f4f7e5ec7c7c9974e3))
 * Bezier curves. ([de20f9e](https://github.com/NazioLT/NZCore/commit/de20f9e6da5267583d03c9d35c55a0132c302175))
 * Bezier Derivative. ([f60ab1d](https://github.com/NazioLT/NZCore/commit/f60ab1dc6894555bf036b584d3be0cc8d2a31ae6))
-* Bezier 2nd Derivative. ([]())
+* Bezier 2nd Derivative. ([dc738de](https://github.com/NazioLT/NZCore/commit/dc738de682a9671853509366bfbd0f6d0ef45325))
 
 #### Utils Library